fix: cast item_count to integer — prevents string concatenation in stat cards
PostgreSQL COUNT(*) returns bigint, serialized as string by Neon driver.
"0" is truthy so `|| 0` fallback doesn't trigger, causing reduce() to
concatenate strings (e.g., "0000000000" instead of 0).

Fix: Add Number() casts in UI reduce/display and ::int casts in SQL queries.

+++ pages/api/procurement/categories/[id].ts
@@ -47,8 +47,8 @@ async function handleGet(id: string, res: NextApiResponse) {
       sc.*,
       pc.name as parent_name,
       pc.code as parent_code,
-      (SELECT COUNT(*) FROM stock_items WHERE category_id = sc.id) as item_count,
-      (SELECT COUNT(*) FROM stock_categories WHERE parent_id = sc.id) as child_count
+      (SELECT COUNT(*)::int FROM stock_items WHERE category_id = sc.id) as item_count,
+      (SELECT COUNT(*)::int FROM stock_categories WHERE parent_id = sc.id) as child_count
     FROM stock_categories sc
     LEFT JOIN stock_categories pc ON sc.parent_id = pc.id
     WHERE sc.id = ${id}
